+++ frontend/src/components/AppFooter/AppFooter.module.scss
@@ -4,7 +4,6 @@
   display: grid;
   @include container;
   padding: 0;
-  margin-top: -12px;
   p {
     padding: 10px;
     font-size: 14px;
